feat: ⬆️ updating to react-scripts 5+
react-scripts 5 is now using WebPack5 in which there is no polyfills
by default anymore.

There is a PR on their repo about adding them back.
For now I propose to use craco, which is a way to extend react-scripts
without ejecting it. This allows us to extend the webpack config
and add polyfills.