feat: 🫵 token-details-ui (#192)
* fix: ⚡️ kick the PR

Changing the flex direction

* fix: ⚡️ use classes

* fix: ⚡️ ues memo for tabs

* fix: ⚡️ dens feedback

* fix: ⚡️ checkbox colors

* fix: 🎨 small ajustment

* Merge branch 'develop' into mel/develop/token-details-ui

* fix: ⚡️ slink edits (#193)

- changed colours
- removed curation

* fix: ⚡️ re-add objkt number

* fix: ⚡️ use-icon for fullscreen button

* fix: ⚡️ fullscreen icon (wip)

* feat: 🔥 add mint date to the infos list

* fix: ⚡️ add the missing lists to use-settings

* fix: ⚡️ make use of the lists

* fix: ⚡️ rename list to map

* fix: ⚡️ remove HEN lists

what was needed has been forked.

* fix: ⚡️ use nsfw in token-masonry

@xat you might want to check this commit
Not sure if it's the best approach now.

* fix: ⚡️ match the proxy limit (2gb for now)

* fix: ⚡️ figma consistency (#195)

Co-authored-by: Mel Massadian <melmassadian@gmail.com>

* feat: 🔥 margins

* fix: ⚡️ color & margins

* fix: ⚡️ re-order containers/padding

* fix: 🔥 ipfs color

* fix: ⚡️ basic info spacing

* fix: 🎨 align to tabs

* fix: 🎨 reduce margins

* fix: ⚡️ inline styles

* feat: 🇵🇰 pakistan feed (#196)

Heavily edited cherry pick, the following is not super accurate now:

* feat: ⚡️ template for the upcoming contract
* fix: ✨ use aria key
* fix: ⚡️ use the real address
* fix: ⚡️ ban filter

This is solved in develop, there was no
filtering on half the feeds...

* feat: ⚡️ store the current tab in states

The `prev` state was unused.
I added a `current` one that simply stores the value of the current
tag.
This is then used to display a simple notice on the fund feeds for
instance.

* fix: ⚡️ mobile style & new text
* fix: ⚡️ text layout
* fix: ⚡️ fix banner
* fix: ⚡️ infinite scroll
* fix: ⚡️ fetch 20 per batch

* fix: 🎨 pointer events issue on vector tokens (#185)

* fix: banner obscuring restricted notice

* fix: ⚡️ duplicates in fund feeds (#199)

Cherry picked and adapted to develop.
i.e the query in each fund feed, the one in hicdex.js is not
used anywhere anymore.

* fix: ⚡️ duplicates in fund feeds
* fix: 🔥 move filters down

* fix: 🐛 disable pointer events on feeds (SVG) (#206)

* feat: 🚀 add Tezos for Iran (#207)

* feat: 🚀 add Tezos for Iran

- Separated tag row from fundraise row.
- Replaced a tags with buttons (not component yet)

* fix: ⚡️ add tag as well

* fix: 🎨 add multiple extra tags

* feat: ⚡️ feed block list

A pending PR to teia-report will add the appropriate file for this.
TEMP SOLUTION TO ADAPT IN DEVELOP

* fix: 🎨 typo

Also the file is now merged in teia-report

* fix: ⚡️ creations | collabs sorting

* fix: 🎨 filter labels on very small screens

* fix: 🎨 from merge

Something is wrong here, too many lists.

* fix: ⚡️ remove unused lists (HEN, OBJKT etc..)

I made a PR in report to reintroduce
a per objkt ban, but AFAIK it was
not used anymore in HEN's code
anyway..

Co-authored-by: Spencer Linkous <83504509+SlinkousArt@users.noreply.github.com>
Co-authored-by: NoRulesJustFeels <entertailion@gmail.com>