fix: 🔥 comment out local settings rm functions + some types
These are not yet used and I don't see any use for them yet for our
use cases.

Also ordered the keys from the request and the types to match.
I want to separately type the different stages and NFT object can have
so that we properly know which are available when.