fix: ✨ hacky way to fix custom rpc input
Ideally the whole settings should be a form that supports validation.
We have the component but it should be a separate PR imo.
The other option I added a todo for is having custom validation / input
which should be quite easy to add to the existing one, it should be
done from within onBlur while still supporting a custom onBlur.